Start implementation of POST and GET. Separating out the functionality of each.
As of this commit, the libbson library doesn't support parsing JSON into
BSON. Just BSON into JSON. We need to either 1) manually parse the JSON
into BSON or 2) Utilize some kind of library.

And by 1), it looks like it might be useful to use the MDBWrapper's
json_to_bson parsing. This is just a simple function to take a json
object, cycle through it by key value and append those key/value pairs
to a corresponding BSON object. Not too crazy.